requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.11.3
bokeh==0.11.0
dill==0.2.4
gunicorn==19.4.5
itsdangerous==0.24
natsort==4.0.4
numpy==1.10.4
pandas==0.17.1
pydicom==0.9.9
python-dateutil==2.4.2
pytz==2015.7
requests==2.9.1
six==1.10.0
tornado==4.3
ujson==1.34
